Move ToolStripOverflowAccessibleObject to a separate file
